task: store, retrieve, update favTokens from MongoDb
